requirements.txt
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
greenlet==3.0.3
idna==3.7
kaggle==1.6.12
numpy==1.26.4
opendatasets==0.1.22
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.29
text-unidecode==1.3
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
